added conf_editor to site system
