+++ index.php
@@ -29,14 +29,13 @@ if(! file_exists($fpath)) {
             case "help_details":
             case "help_using":
             case "website_editor":
+            case "conf_editor":
                 $site_class = $_GET["p"];
                 break;
 
             case "ma_editor":
             case "remove_caches":
-            case "conf_editor":
             case "design_editor":
-            case "areas_manager":
             case "update":
             case "logout":
                 header('Location: ' . $_GET["p"] . ".php");
